'email_body' column name was changed to 'message_id' in email table in mysql database
Posting job applicatoin end point working.

Fecthing and storing emails of the company provided into the database is working!
