+++ seed_data/emails.js
@@ -3,7 +3,7 @@ module.exports = [
     id: 1,
     id_of_company: 1,
     date: "234234234",
-    email_body: "<p>this is email body</p>",
+    message_id: "12345678",
     subject: "Whiteboarding",
   },
 ];
